net: phy: micrel: Allow the PHY to get PTP clock externally
The LAN8814 PTP timestamping unit requires a 125 MHz reference clock.
This can be derived from an internal source or an external reference
clock.

Read the optional 'microchip,ptp-clock-frequency-hz' property from the
device tree. If this property is present, it indicates that an external
reference clock is used.

In this case, configure the PTP Reference Clock Configuration register
to select the external source and program the PLL dividers (DIVR and
DIVF) to synthesize the required 125 MHz PTP clock based on the
provided frequency (10MHz, 25MHz, or 125MHz).

If the property is absent, the driver continues to use the default
internal clock source.

Signed-off-by: Horatiu Vultur <horatiu.vultur@microchip.com>