[DNM] net: lan966x: Register recovered clock pin
Look for 'dpll-pins' property inside firmware node and if it is present
and associated dpll pin is already registered then register own pin-on-pin
dpll pin and use it as recovered clock pin for the network device.

If the dpll pin referenced by 'dpll-pins' property is not registered yet
then wait for its registration.

To do so subscribe for DPLL notifications and check if the referenced
external dpll pin is registered then perform own pin registration above.
If the referenced dpll pin becomes unregistered then unregister also
own pin.

Signed-off-by: Ivan Vecera <ivecera@redhat.com>